kapruka-concierge/requirements.txt
playwright
qdrant-client
python-dotenv
openai
fastapi
uvicorn[standard]
python-multipart
filelock
rapidfuzz
pytest
httpx
aiofiles
pillow
